Recipes for migrating common 'get' methods for Joda-Time to java.time
Closes SEN-2541

This set of recipes covers the migration of some methods that are common
to a range of java.time classes.

Tests that demonstrate and support the logic of the recipes are included
in CommonGetMethodsTest.

Note: the DateTime getMillis recipe is also included, I originally
thought this would be a common method but it is really only for DateTime
and it's corresponding java.time objects. I have included it anyway but
it is in the DateTime folder to reflect it is only for DateTime migration.